requirements/development.txt
psycopg2>=2.6.2
mock>=2.0.0

dj-database-url>=0.4.1
pytest>=3.1.3
pytest-django>=3.1.2
pytest-cov>=2.5.1
pytest-sugar>=0.8.0
pytest-watch>=4.1.0
pytest-env>=0.6.2
